# data/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Contact Management API By Pavan",
    "description": "API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "description": "Authenticates a user and returns a JWT token store as in form of token under token name.",
        "operationId": "loginUser",

        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the user."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "description": "The password for the user account."
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },

        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "string",
                      "example": "Login successful"
                    }
                  }
                }
              }
            }
          },

          "401": {
            "description": "Invalid Credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid Credentials"
                    }
                  }
                }
              }
            }
          },

          "400": {
            "description": "Invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Invalid fields" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/emailverify": {
      "get": {
        "summary": "Verify user email",
        "description": "Verifies the user's email using a verification token.",
        "operationId": "verifyEmail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "The verification token sent to the user's email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "string",
                      "example": "Email verified Successfully"
                    }
                  }
                }
              }
            }
          },
          "400_Expired_Token": {
            "description": "Token is expired and new OTP link generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Token is expired and new OTP link generated"
                    }
                  }
                }
              }
            }
          },

          "400_Token_Not_Found": {
            "description": "Token is missing in db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Token is missing in db"
                    }
                  }
                }
              }
            }
          },

          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/newpassword": {
      "post": {
        "summary": "Reset user password",
        "description": "Resets the user's password using a valid reset token.",
        "operationId": "resetUserPassword",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "The password reset token sent to the user's email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPasswordRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "string",
                      "example": "Password Reset Completed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (validation error)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404_Invalid_Token": {
              "description": "'Invalid Token' or 'Email not existing in db'",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "'Invalid Token'"
                      }
                    }
                  }
                }
              }
            },

            "404_Expired_Token": {
              "description": "'Token is expired'",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "'Token is expired'"
                      }
                    }
                  }
                }
              }
            },

            "500": {
              "description": "'Internal Server Error.'",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "name": "token",
        "in": "header",
        "description": "The authorization token in the form of `Bearer <token>`"
      }
    },
...
    "schemas": {
...
      "NewPasswordRequestBody": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": "6",
            "description": "The new password for the user account."
          }
        },
        "required": ["password"]
      }
    },
...
  }
}